// HFY.MvcWebUI.Angular/ClientApp/src/app/lazy-general-page/classes/lastikler/models/lastik-kopyala.ts
export class LastikKopyala {
    AsilLastikID:number;

    FirmaID: number;

    LastikMarkaID: number;

    LastikMarkaDesenID: number;

    SeriNo: string;

    Fiyat: number;

    EbatID: number;

    LastikTurID: number;

    KayitTarihi:string;

    LastikTipID: number;

    DisSeviyesi: number;

    LastikKilometre: number;

    LastikKonumID: number;

}